+++ src/lib/Components/Table/Table.test.tsx
@@ -26,27 +26,3 @@ it('renders without crashing', () => {
   ReactDOM.render(<Table {...tableProps} />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
-
-it('should not dispatch ScrollTable on scroll', () => {
-  const wrapper = mount(<Table {...tableProps} />);
-  expect(wrapper.find('.ka-table-wrapper').prop('onScroll')).toBeUndefined();
-  expect(tableProps.dispatch).toBeCalledTimes(0);
-});
-
-it('should dispatch ScrollTable on scroll in case of virtual scrolling', () => {
-  const wrapper = mount(
-    (
-      <Table {...tableProps} virtualScrolling={({
-        itemHeight: 10,
-        tbodyHeight: 100
-      })}/>
-    )
-  );
-  const scrollTop = 11;
-
-  wrapper.find('.ka-table-wrapper').prop('onScroll')!({ currentTarget: {scrollTop} } as any);
-  expect(tableProps.dispatch).toBeCalledTimes(1);
-  expect(tableProps.dispatch).toBeCalledWith(
-    { type: ActionType.ScrollTable, scrollTop },
-  );
-});
